fix tools-host can't download error

+++ .github/scripts/steps/build.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-make || make 
-rm -f images/latest.image
+make && rm -f images/latest.image
